Add support root-level `Style` element handling
