+++ Sources/BuiltinRegistryGenerator/BuiltinRegistryGenerator.swift
@@ -42,7 +42,6 @@ struct BuiltinRegistryGenerator: ParsableCommand {
         "RenameButton": "RenameButton()",
         "RoundedRectangle": "Shape(shape: RoundedRectangle(from: element))",
         "Color": "ColorView()",
-        "compiled-lvn-stylesheet": "CompiledLVNStylesheet<R>()",
         "phx-main": "PhxMain<R>()",
     ]
